+++ check_exports.js
@@ -1,2 +1,3 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
 const panels = require('react-resizable-panels');
 console.log(Object.keys(panels));